feat: #tno-1730 - added basic filter import/export - changes as per feedback
