+++ app/editor/src/features/admin/filters/utils/FilterImportExportModel.ts
@@ -1,0 +1,7 @@
+import { IFilterImportExportModel } from './IFilterImportExportModel';
+
+export class FilterImportExportModel implements IFilterImportExportModel {
+  name?: string | undefined;
+  description?: string | undefined;
+  settings: any;
+}
