+++ app/editor/src/features/admin/filters/utils/IFilterImportExportModel.ts
@@ -1,0 +1,5 @@
+export interface IFilterImportExportModel {
+  name?: string;
+  description?: string;
+  settings: any;
+}
