+++ app/editor/src/features/admin/filters/utils/index.ts
@@ -1,4 +1,7 @@
-export * from './filterImportExport';
+export * from './exportFilter';
+export * from './FilterImportExportModel';
 export * from './generateMultiMatch';
 export * from './getActionOptions';
 export * from './getTagOptions';
+export * from './IFilterImportExportModel';
+export * from './parseExportedFilter';
